requirements.txt
flask
BeautifulSoup4
micawber
pygments
markdown
peewee
flask-sqlalchemy
Flask-Script
Flask-Migrate
Flask-Shell